@@ -1,40 +1,38 @@
+# frozen_string_literal: true
+# TODO: replace with ruby-progressbar dependency
 module Bumbler
   module Progress
     @item_count   = 0
     @loaded_items = 0
 
     # registry[item_type][item_name] = {:time => 123.45}
-    @registry = Hash.new { |h,k| h[k] = {} }
+    @registry = Hash.new { |h, k| h[k] = {} }
 
     class << self
-      def registry
-        @registry
-      end
+      attr_reader :registry
 
       def register_item(type, name)
         # Build a blank key for the item
-        unless @registry[type][name]
-          @item_count  += 1
-        end
+        @item_count += 1 unless @registry[type][name]
 
         @registry[type][name] = {}
       end
 
       def item_started(type, name)
-        @curr_item = {:type => type, :name => name}
+        @curr_item = { type: type, name: name }
 
-        self.render_progress
+        render_progress
       end
 
       def item_finished(type, name, time)
-        @registry[type][name] = {:time => time}
+        @registry[type][name] = { time: time }
 
-        @loaded_items  += 1
+        @loaded_items += 1
 
-        @prev_item = {:type => type, :name => name, :time => time}
+        @prev_item = { type: type, name: name, time: time }
         @curr_item = nil if @curr_item && @curr_item[:name] == @prev_item[:name] && @curr_item[:type] == @prev_item[:type]
 
-        self.render_progress
+        render_progress
       end
 
       def start!
@@ -48,11 +46,11 @@ module Bumbler
       def bar(width)
         inner_size = width - 2
 
-        fill_size = [((@loaded_items.to_f / @item_count.to_f) * inner_size).to_i, inner_size].min
-        fill  = '#' * fill_size
+        fill_size = [((@loaded_items / @item_count.to_f) * inner_size).to_i, inner_size].min
+        fill = '#' * fill_size
         empty = ' ' * (inner_size - fill_size)
 
-        return "[#{fill}#{empty}]"
+        "[#{fill}#{empty}]"
       end
 
       def render_progress
@@ -64,12 +62,12 @@ module Bumbler
         #   (##/##) <current>...   <prev> (####.##ms)
         #
         # Skip the current if there isn't enough room
-        count   = '(%s/%d) ' % [@loaded_items.to_s.rjust(@item_count.to_s.size), @item_count]
+        count   = format('(%s/%d) ', @loaded_items.to_s.rjust(@item_count.to_s.size), @item_count)
         current = @curr_item ? "#{@curr_item[:name]}... " : ''
-        prev    = @prev_item ? '%s (%sms)' % [@prev_item[:name], ('%.2f' % @prev_item[:time]).rjust(7)] : ''
+        prev    = @prev_item ? format('%s (%sms)', @prev_item[:name], ('%.2f' % @prev_item[:time]).rjust(7)) : ''
 
         if $stdout.tty?
-          width = self.tty_width
+          width = tty_width
 
           print "\r\e[A\r\e[A" if @outputted_once
           @outputted_once = true
@@ -78,7 +76,7 @@ module Bumbler
           space_for_current = width - (count.length + prev.length)
 
           # Render the progress
-          puts self.bar(width)
+          puts bar(width)
 
           if space_for_current >= current.length
             puts count + current + prev.rjust(width - count.length - current.length)
@@ -86,7 +84,7 @@ module Bumbler
             puts count + prev.rjust(width - count.length)
           end
         elsif @curr_item
-          puts '%s %s' % [count, @curr_item[:name]]
+          puts format('%s %s', count, @curr_item[:name])
         end
       end
     end
