@@ -1,54 +1,61 @@
+# frozen_string_literal: true
 module Bumbler
   module Bundler
     class << self
       # Returns which gem a require maps to, or nil.
       def gem_for_require(path)
-        self.read_bundler_environment if @require_map.nil?
-        
-        return @require_map[path]
+        read_bundler_environment if @require_map.nil?
+
+        @require_map[path]
       end
-      
+
       def require_started(path)
-        gem_name = self.gem_for_require(path)
+        gem_name = gem_for_require(path)
         return unless gem_name
-        
+
         Bumbler::Progress.item_started(:bundler, gem_name)
       end
-      
+
       def require_finished(path, load_time)
-        self.read_bundler_environment if @gem_state.nil?
-        
+        read_bundler_environment if @gem_state.nil?
+
         # Tick it off for the gem.
-        gem_name = self.gem_for_require(path)
+        gem_name = gem_for_require(path)
         return unless gem_name
-        
+
         @gem_state[gem_name][path] = true
-        
-        if @gem_state[gem_name].values.all?
-          Bumbler::Progress.item_finished(:bundler, gem_name, load_time)
-        end
+
+        Bumbler::Progress.item_finished(:bundler, gem_name, load_time) if @gem_state[gem_name].values.all?
       end
-      
+
       def start!
-        self.read_bundler_environment
+        read_bundler_environment
       end
-      
+
+      private
+
       def read_bundler_environment
         @require_map = {}
         @gem_state = {}
-        
+
         ::Bundler.environment.current_dependencies.each do |spec|
-          @gem_state[spec.name] = {}
-          
-          Array(spec.autorequire || spec.name).each do |path|
-            # Handle explicitly required gems
-            path = spec.name if path == true
-            
-            @require_map[path] = spec.name
-            @gem_state[spec.name][path] = false
+          gem_name = spec.name
+          @gem_state[gem_name] = {}
+
+          # TODO: this is horrible guess-work ... we need to get the gens load-path instead
+          paths =
+            if !spec.autorequire || spec.autorequire == [true]
+              [gem_name]
+            else
+              spec.autorequire
+            end
+
+          paths.each do |path|
+            @require_map[path] = gem_name
+            @gem_state[gem_name][path] = false
           end
-          
-          Bumbler::Progress.register_item(:bundler, spec.name)
+
+          Bumbler::Progress.register_item(:bundler, gem_name)
         end
       end
     end
