@@ -1,20 +1,16 @@
+# frozen_string_literal: true
 require 'bumbler'
+require 'bundler'
 
-# Do nothing unless we're in a bundle
-begin
-  require 'bundler'
-  # This raises if there isn't a gemfile in our root
-  Bundler.default_gemfile
-  # Workaround for Ruby 2.3, see: github.com/nevir/Bumbler/issues/12
-  Bumbler::Bundler
+# This raises if there isn't a gemfile in our root
+Bundler.default_gemfile
 
-  # Kick it off
-  Bumbler::Hooks.hook_require!
-  Bumbler::Hooks.watch_require!
+# Workaround for Ruby 2.3, see: github.com/nevir/Bumbler/issues/12
+Bumbler::Bundler.name
 
-  Bumbler::Bundler.start!
-  Bumbler::Progress.start!
+# Kick it off
+Bumbler::Hooks.hook_require!
+Bumbler::Hooks.watch_require!
 
-rescue
-  # Welp, if we fail, we fail.
-end
+Bumbler::Bundler.start!
+Bumbler::Progress.start!
