@@ -1,32 +1,38 @@
+# frozen_string_literal: true
 require_relative "test_helper"
 require "tmpdir"
 
 describe Bumbler do
-  def sh(command, options={})
-    result = Bundler.with_clean_env { `#{command} #{"2>&1" unless options[:keep_output]}` }
-    raise "#{options[:fail] ? "SUCCESS" : "FAIL"} #{command}\n#{result}" if $?.success? == !!options[:fail]
+  def sh(command, fail: false, keep_output: false)
+    result = Bundler.with_clean_env { `#{command} #{"2>&1" unless keep_output}` }
+    raise "#{fail ? "SUCCESS" : "FAIL"} #{command}\n#{result}" if $?.success? == fail
     result
+  end
+
+  def write(path, content)
+    FileUtils.mkdir_p(File.dirname(path))
+    File.write(path, content)
   end
 
   around { |test| Dir.mktmpdir { |dir| Dir.chdir(dir) { test.call } } }
 
   it "prints simple progress without tty on ruby project" do
-    File.write("Gemfile", "source 'https://rubygems.org'\ngem 'fakegem', path: '#{Bundler.root}/test/fakegem'\ngem 'bumbler', path: '#{Bundler.root}'")
-    File.write("test.rb", "require 'bumbler/go'\nBundler.require")
+    write("Gemfile", "source 'https://rubygems.org'\ngem 'fakegem', path: '#{Bundler.root}/test/fakegem'\ngem 'bumbler', path: '#{Bundler.root}'")
+    write("test.rb", "require 'bumbler/go'\nBundler.require")
     result = sh "bundle exec ruby test.rb"
     result.strip.must_equal "(0/2)  fakegem\n(1/2)  bumbler"
   end
-  
+
   it "includes gems that are explicitly required" do
-    File.write("Gemfile", "source 'https://rubygems.org'\ngem 'fakegem', path: '#{Bundler.root}/test/fakegem', require: true\ngem 'bumbler', path: '#{Bundler.root}'")
-    File.write("test.rb", "require 'bumbler/go'\nBundler.require")
+    write("Gemfile", "source 'https://rubygems.org'\ngem 'fakegem', path: '#{Bundler.root}/test/fakegem', require: true\ngem 'bumbler', path: '#{Bundler.root}'")
+    write("test.rb", "require 'bumbler/go'\nBundler.require")
     result = sh "bundle exec ruby test.rb"
     result.strip.must_equal "(0/2)  fakegem\n(1/2)  bumbler"
   end
 
   describe "CLI" do
-    def bumbler(command="", options={})
-      sh("#{Bundler.root}/bin/bumbler #{command}", options)
+    def bumbler(command = "", **args)
+      sh("#{Bundler.root}/bin/bumbler #{command}", **args)
     end
 
     it "shows --version" do
@@ -37,30 +43,42 @@ describe Bumbler do
       bumbler("bumbler --help").must_include "Bumbler"
     end
 
+    it "fails when given arguments" do
+      bumbler("bumbler help", fail: true).must_include "arguments"
+    end
+
+    it "fails without Gemfile" do
+      bumbler("", fail: true).must_include "Could not locate Gemfile"
+    end
+
+    it "fails without config/environment.rb" do
+      write("Gemfile", "source 'https://rubygems.org'")
+      bumbler("", fail: true).must_include "./config/environment"
+    end
+
     describe "with simple gemfile" do
-      def structure
-        File.write("Gemfile", "source 'https://rubygems.org'\ngem 'fakegem', path: '#{Bundler.root}/test/fakegem'")
-        FileUtils.mkdir_p("config")
-        File.write("config/environment.rb", "require 'bundler/setup'\nrequire 'fakegem'")
+      before do
+        write("Gemfile", "source 'https://rubygems.org'\ngem 'fakegem', path: '#{Bundler.root}/test/fakegem'")
+        write("config/environment.rb", "require 'bundler/setup'\nrequire 'fakegem'")
       end
 
       it "prints simple progress without tty" do
-        structure
         bumbler.strip.must_equal "(0/1)  fakegem\nSlow requires:"
       end
 
+      it "can show all" do
+        bumbler("--all").strip.must_match(/^\(0\/1\)\s+fakegem\s+Stats for bundler items:\s+\d+\.\d+\s+fakegem$/m)
+      end
+
       it "shows more with lower threshold" do
-        structure
-        bumbler("-t 0").strip.must_match /^Slow requires:\s+\d+\.\d+\s+fakegem$/m
+        bumbler("-t 0").strip.must_match(/^Slow requires:\s+\d+\.\d+\s+fakegem$/m)
       end
     end
 
     describe "with initializers" do
       it "records initializers" do
-        File.write("Gemfile", "")
-        FileUtils.mkdir_p("config")
-
-        File.write("config/application.rb", <<-RUBY)
+        write("Gemfile", "")
+        write("config/application.rb", <<-RUBY)
           $offset = 0.0
           def wait
             sleep 0.1 + $offset
@@ -96,7 +114,7 @@ describe Bumbler do
           end
         RUBY
 
-        File.write("config/environment.rb", <<-RUBY)
+        write("config/environment.rb", <<-RUBY)
           Rails::Engine.new.load(Rails.root + "/config/initializers/foo.rb")
           Rails::Engine.new.load(Rails.root + "/config/initializers/bar.rb")
           Rails::Initializable::Initializer.new(:baz).run(1,2,3)
