@@ -1,4 +1,6 @@
+# frozen_string_literal: true
 require 'bundler/setup'
+require 'maxitest/global_must'
 require 'maxitest/autorun'
 
-$LOAD_PATH.unshift File.expand_path("../../lib", __FILE__)
+$LOAD_PATH.unshift File.expand_path('../lib', __dir__)
